feat(backend): Implementado gerenciamento de produtos e categorias com upload de imagens

+++ src/main/java/com/tcc/serveme/api/repository/GenericRepository.java
@@ -8,7 +8,7 @@ public interface GenericRepository<T, ID> {
 
     List<T> findAll();
 
-    int save(T entity);
+    long save(T entity);
 
     int update(T entity);
